#82 refactor: create deep session module, fix bigint types, add resolved_state
- Fix all i64/u64 → number in ts-rs annotations across all Rust types
- Add resolved_state to SessionEventPayload (eliminates RUN_STATE_MAP)
- Create src/lib/modules/sessions/ merging context + commands + types
- Move sessionsByIssueId derivation into module
- Move event listeners into module (setup in setSessionsContext)
- Delete sessions.context.svelte.ts, session_commands.ts, types/session.ts
