+++ src-tauri/src/models/github.rs
@@ -62,9 +62,11 @@ pub struct GitHubStatusCache {
     pub issue_id: String,
     pub branch_status: Option<String>,
     pub pr_state: Option<PullRequestState>,
+    #[ts(type = "number | null")]
     pub pr_number: Option<i64>,
     pub pr_url: Option<String>,
     pub github_issue_state: Option<String>,
+    #[ts(type = "number | null")]
     pub behind_base_count: Option<i64>,
     pub merge_conflict: Option<bool>,
     pub fetched_at: Option<String>,
@@ -84,6 +86,7 @@ pub enum GhCliAvailability {
 #[derive(Debug, Clone, Serialize, Deserialize, TS)]
 #[ts(export)]
 pub struct AssignedIssue {
+    #[ts(type = "number")]
     pub number: i64,
     pub title: String,
     pub state: String,
