+++ src-tauri/src/models/issue.rs
@@ -15,6 +15,7 @@ pub struct Issue {
     pub color: Option<String>,
     pub status: String,
     pub github_issue_url: Option<String>,
+    #[ts(type = "number | null")]
     pub github_issue_number: Option<i64>,
     pub branch_name: Option<String>,
     pub base_branch: Option<String>,
@@ -23,10 +24,13 @@ pub struct Issue {
     pub parent_issue_id: Option<String>,
     pub editor_folder: Option<String>,
     pub dev_server_command: Option<String>,
+    #[ts(type = "number | null")]
     pub dev_server_port: Option<i64>,
+    #[ts(type = "number | null")]
     pub dev_server_pid: Option<i64>,
     pub browser_url: Option<String>,
     pub labels: Option<String>,
+    #[ts(type = "number")]
     pub sort_order: i64,
     pub created_at: String,
 }
